Move font dependent text helpers out of StringUtils
